pyproject.toml
[tool.poetry]
name = "experience_player"
version = "0.1.0"
description = ""
authors = ["frapercan <frapercan1@alum.us.es>"]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
torch = "^2.0.1"
torchvision = "^0.15.2"
pyautogui = "^0.9.54"
matplotlib = "^3.7.1"
pynput = "^1.7.6"


[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
coverage = {extras = ["toml"], version = "^7.2.1"}
flake8 = "^6.0.0"
flake8-bugbear = "^23.2.13"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.10.3"

[tool.coverage.run]
omit = ["*/site-packages/*"]

[tool.coverage.report]
fail_under = 100

[tool.taskipy.tasks]
html_docs = "make html -C docs"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

